clean: use action names and selectors
